+++ client/src/App.css
@@ -8,55 +8,6 @@
   align-items: center;
 }
 
-.form-search {
-  margin: 0 auto;
-  margin-top: -20%;
-  text-align: center;
-  font: bold 20px sans-serif;
-  max-width: 70%;
-  width: 80%;
-  position: relative;
-}
-
-.form-search input {
-  width: 60%;
-  box-sizing: border-box;
-  font-size: 20px;
-  border-bottom-left-radius: 2px;
-  border-top-left-radius: 2px;
-  background-color: #ffffff;
-  box-shadow: 1px 2px 4px 0 rgba(0, 0, 0, 0.08);
-  padding: 14px 15px 14px 40px;
-  border: 1px solid #b6c3cd;
-  border-right: 0;
-  color: #4E565C;
-  outline: none;
-  -webkit-appearance: none;
-}
-
-.form-search button {
-  border-bottom-right-radius: 2px;
-  border-top-right-radius: 2px;
-  background-color: #6caee0;
-  box-shadow: 1px 2px 4px 0 rgba(0, 0, 0, 0.08);
-  font-size: 20px;
-  color: #ffffff;
-  padding: 15px 22px;
-  margin-left: -4px;
-  cursor: pointer;
-  border: none;
-  outline: none;
-}
-
-.form-search i {
-  position: relative;
-  top: 15px;
-  left: 20px;
-  font-size: 16px;
-  color: #80A3BD;
-}
-
-
 .App-logo {
   height: 40vmin;
   pointer-events: none;
